+++ sdks/java/io/snowflake/src/test/java/org/apache/beam/sdk/io/snowflake/test/unit/write/CreateDispositionTest.java
@@ -17,16 +17,19 @@
  */
 package org.apache.beam.sdk.io.snowflake.test.unit.write;
 
+import static org.apache.beam.sdk.io.snowflake.test.TestUtils.getCsvMapper;
 import static org.junit.Assert.assertTrue;
 
 import java.sql.SQLException;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.LongStream;
 import org.apache.beam.sdk.io.snowflake.SnowflakeIO;
 import org.apache.beam.sdk.io.snowflake.SnowflakePipelineOptions;
-import org.apache.beam.sdk.io.snowflake.enums.WriteDisposition;
+import org.apache.beam.sdk.io.snowflake.data.SnowflakeColumn;
+import org.apache.beam.sdk.io.snowflake.data.SnowflakeTableSchema;
+import org.apache.beam.sdk.io.snowflake.data.text.SnowflakeVarchar;
+import org.apache.beam.sdk.io.snowflake.enums.CreateDisposition;
 import org.apache.beam.sdk.io.snowflake.services.SnowflakeService;
 import org.apache.beam.sdk.io.snowflake.test.FakeSnowflakeBasicDataSource;
 import org.apache.beam.sdk.io.snowflake.test.FakeSnowflakeDatabase;
@@ -45,15 +48,17 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 @RunWith(JUnit4.class)
-public class QueryDispositionLocationTest {
+public class CreateDispositionTest {
   private static final String FAKE_TABLE = "FAKE_TABLE";
-  private static final String BUCKET_NAME = "BUCKET";
+  private static final String BUCKET_NAME = "bucket/";
 
   @Rule public final transient TestPipeline pipeline = TestPipeline.create();
   @Rule public ExpectedException exceptionRule = ExpectedException.none();
 
   private static SnowflakePipelineOptions options;
   private static SnowflakeIO.DataSourceConfiguration dc;
+  private static String stagingBucketName;
+  private static String storageIntegrationName;
 
   private static SnowflakeService snowflakeService;
   private static List<Long> testData;
@@ -62,43 +67,45 @@ public class QueryDispositionLocationTest {
   public static void setupAll() {
     PipelineOptionsFactory.register(SnowflakePipelineOptions.class);
     options = TestPipeline.testingPipelineOptions().as(SnowflakePipelineOptions.class);
+    options.setStagingBucketName(BUCKET_NAME);
+    options.setServerName("NULL.snowflakecomputing.com");
+
+    stagingBucketName = options.getStagingBucketName();
+    storageIntegrationName = options.getStorageIntegrationName();
 
     snowflakeService = new FakeSnowflakeServiceImpl();
     testData = LongStream.range(0, 100).boxed().collect(Collectors.toList());
-  }
-
-  @Before
-  public void setup() {
-    options.setStagingBucketName(BUCKET_NAME);
-    options.setStorageIntegrationName("STORAGE_INTEGRATION");
-    options.setServerName("NULL.snowflakecomputing.com");
 
     dc =
         SnowflakeIO.DataSourceConfiguration.create(new FakeSnowflakeBasicDataSource())
             .withServerName(options.getServerName());
   }
 
+  @Before
+  public void setup() {}
+
   @After
   public void tearDown() {
     TestUtils.removeTempDir(BUCKET_NAME);
+    FakeSnowflakeDatabase.clean();
   }
 
   @Test
-  public void writeWithWriteTruncateDispositionSuccess() throws SQLException {
+  public void writeWithWriteCreateDispositionWithAlreadyCreatedTableSuccess() throws SQLException {
     FakeSnowflakeDatabase.createTable(FAKE_TABLE);
 
     pipeline
         .apply(Create.of(testData))
         .apply(
-            "Truncate before write",
+            "Copy IO",
             SnowflakeIO.<Long>write()
                 .withDataSourceConfiguration(dc)
                 .to(FAKE_TABLE)
-                .withStagingBucketName(options.getStagingBucketName())
-                .withStorageIntegrationName(options.getStorageIntegrationName())
+                .withStagingBucketName(stagingBucketName)
+                .withStorageIntegrationName(storageIntegrationName)
                 .withUserDataMapper(TestUtils.getLongCsvMapper())
                 .withFileNameTemplate("output")
-                .withWriteDisposition(WriteDisposition.TRUNCATE)
+                .withCreateDisposition(CreateDisposition.CREATE_IF_NEEDED)
                 .withSnowflakeService(snowflakeService));
 
     pipeline.run(options).waitUntilFinish();
@@ -109,52 +116,100 @@ public class QueryDispositionLocationTest {
   }
 
   @Test
-  public void writeWithWriteEmptyDispositionWithNotEmptyTableFails() {
-    FakeSnowflakeDatabase.createTableWithElements(FAKE_TABLE, Arrays.asList("NOT_EMPTY"));
+  public void writeWithWriteCreateDispositionWithCreatedTableWithoutSchemaFails() {
 
     exceptionRule.expect(RuntimeException.class);
     exceptionRule.expectMessage(
-        "java.lang.RuntimeException: Table is not empty. Aborting COPY with disposition EMPTY");
+        "The CREATE_IF_NEEDED disposition requires schema if table doesn't exists");
 
     pipeline
         .apply(Create.of(testData))
         .apply(
-            "Write SnowflakeIO",
+            "Copy IO",
             SnowflakeIO.<Long>write()
                 .withDataSourceConfiguration(dc)
-                .to(FAKE_TABLE)
-                .withStagingBucketName(options.getStagingBucketName())
-                .withStorageIntegrationName(options.getStorageIntegrationName())
-                .withUserDataMapper(TestUtils.getLongCsvMapper())
+                .to("NO_EXIST_TABLE")
+                .withStagingBucketName(stagingBucketName)
+                .withStorageIntegrationName(storageIntegrationName)
                 .withFileNameTemplate("output")
-                .withWriteDisposition(WriteDisposition.EMPTY)
+                .withUserDataMapper(getCsvMapper())
+                .withCreateDisposition(CreateDisposition.CREATE_IF_NEEDED)
                 .withSnowflakeService(snowflakeService));
 
     pipeline.run(options).waitUntilFinish();
   }
 
   @Test
-  public void writeWithWriteEmptyDispositionWithEmptyTableSuccess() throws SQLException {
+  public void writeWithWriteCreateDispositionWithCreatedTableWithSchemaSuccess()
+      throws SQLException {
+    SnowflakeTableSchema tableSchema =
+        new SnowflakeTableSchema(SnowflakeColumn.of("id", new SnowflakeVarchar()));
+
+    pipeline
+        .apply(Create.of(testData))
+        .apply(
+            "Copy IO",
+            SnowflakeIO.<Long>write()
+                .withDataSourceConfiguration(dc)
+                .to("NO_EXIST_TABLE")
+                .withTableSchema(tableSchema)
+                .withStagingBucketName(stagingBucketName)
+                .withStorageIntegrationName(storageIntegrationName)
+                .withFileNameTemplate("output")
+                .withUserDataMapper(TestUtils.getLongCsvMapper())
+                .withCreateDisposition(CreateDisposition.CREATE_IF_NEEDED)
+                .withSnowflakeService(snowflakeService));
+
+    pipeline.run(options).waitUntilFinish();
+    List<Long> actualData = FakeSnowflakeDatabase.getElementsAsLong("NO_EXIST_TABLE");
+
+    assertTrue(TestUtils.areListsEqual(testData, actualData));
+  }
+
+  @Test
+  public void writeWithWriteCreateDispositionWithCreateNeverSuccess() throws SQLException {
     FakeSnowflakeDatabase.createTable(FAKE_TABLE);
 
     pipeline
         .apply(Create.of(testData))
         .apply(
-            "Write SnowflakeIO",
+            "Copy IO",
             SnowflakeIO.<Long>write()
                 .withDataSourceConfiguration(dc)
                 .to(FAKE_TABLE)
-                .withStagingBucketName(options.getStagingBucketName())
-                .withStorageIntegrationName(options.getStorageIntegrationName())
+                .withStagingBucketName(stagingBucketName)
+                .withStorageIntegrationName(storageIntegrationName)
                 .withFileNameTemplate("output")
                 .withUserDataMapper(TestUtils.getLongCsvMapper())
-                .withWriteDisposition(WriteDisposition.EMPTY)
+                .withCreateDisposition(CreateDisposition.CREATE_NEVER)
                 .withSnowflakeService(snowflakeService));
 
     pipeline.run(options).waitUntilFinish();
-
     List<Long> actualData = FakeSnowflakeDatabase.getElementsAsLong(FAKE_TABLE);
 
     assertTrue(TestUtils.areListsEqual(testData, actualData));
+  }
+
+  @Test
+  public void writeWithWriteCreateDispositionWithCreateNeededFails() {
+
+    exceptionRule.expect(RuntimeException.class);
+    exceptionRule.expectMessage("SQL compilation error: Table does not exist");
+
+    pipeline
+        .apply(Create.of(testData))
+        .apply(
+            "Copy IO",
+            SnowflakeIO.<Long>write()
+                .withDataSourceConfiguration(dc)
+                .to("NO_EXIST_TABLE")
+                .withStagingBucketName(stagingBucketName)
+                .withStorageIntegrationName(storageIntegrationName)
+                .withFileNameTemplate("output")
+                .withUserDataMapper(TestUtils.getLongCsvMapper())
+                .withCreateDisposition(CreateDisposition.CREATE_NEVER)
+                .withSnowflakeService(snowflakeService));
+
+    pipeline.run(options).waitUntilFinish();
   }
 }